fix: renamed the module, fixed the relative imports
